fix warning and bugs
Former-commit-id: 64d692202b3ed3c5ce36b330b4fbc3b9fd3e9c26 [formerly 99ad1314d3d86d5e6b8ea43a4bb4ea83e4740c69] [formerly 531cc13fc157b4774bb9d23fcbb2415e031d23b5 [formerly 83adf7f0dce4d0a28fb2dac08eb93a46a3057326]]
Former-commit-id: 197737c03ce408f1b63f1770a8bd42c821aa1afb [formerly 1df3b7a1a47d4c0e55bcb833b890cd04fb898a61]
Former-commit-id: 6366ae58cd52a53446a7ee80e09285287e7eb311

+++ Frame/Server/Interface/AFINetServerModule.h
@@ -230,7 +230,7 @@ public:
     }
 
 protected:
-    void OnReceiveNetPack(const AFIMsgHead& xHead, const int nMsgID, const char* msg, const uint32_t nLen, const AFGUID& xClientID)
+    void OnReceiveNetPack(const AFIMsgHead& xHead, const int nMsgID, const char* msg, const size_t nLen, const AFGUID& xClientID)
     {
         OnReceiveBaseNetPack(xHead, nMsgID, msg, nLen, xClientID);
     }
@@ -266,7 +266,6 @@ protected:
     }
 
 private:
-
     AFINet* m_pNet;
     int64_t nLastTime;
 };
